making the BearWatcher with modules
