+++ src/javascripts/helpers/data/bearData.js
@@ -1,0 +1,3 @@
+const bearsArr = [];
+
+export default bearsArr;
